Move RewriteRecipeDiscovery to top-level package
