+++ spring-rewrite-commons-docs/src/main/antora/modules/ROOT/pages/partials/example-application-code.adoc
@@ -10,7 +10,7 @@ import org.springframework.rewrite.parser.RewriteProjectParsingResult;
 import org.springframework.rewrite.resource.ProjectResourceSet;
 import org.springframework.rewrite.resource.ProjectResourceSetFactory;
 import org.springframework.rewrite.resource.ProjectResourceSetSerializer;
-import org.springframework.rewrite.recipes.RewriteRecipeDiscovery;
+import org.springframework.rewrite.RewriteRecipeDiscovery;
 import org.springframework.stereotype.Component;
 
 import java.nio.file.Path;
