+++ spring-rewrite-commons-examples/boot-3-upgrade-atomic-example/src/main/java/com/acme/example/SpringBoot3Upgrade.java
@@ -25,7 +25,7 @@ import org.springframework.rewrite.parser.RewriteProjectParsingResult;
 import org.springframework.rewrite.resource.ProjectResourceSet;
 import org.springframework.rewrite.resource.ProjectResourceSetFactory;
 import org.springframework.rewrite.resource.ProjectResourceSetSerializer;
-import org.springframework.rewrite.recipes.RewriteRecipeDiscovery;
+import org.springframework.rewrite.RewriteRecipeDiscovery;
 
 import java.nio.file.Path;
 import java.util.List;
