+++ spring-rewrite-commons-launcher/README.adoc
@@ -98,7 +98,7 @@ import org.springframework.rewrite.parser.RewriteProjectParsingResult;
 import org.springframework.rewrite.resource.ProjectResourceSet;
 import org.springframework.rewrite.resource.ProjectResourceSetFactory;
 import org.springframework.rewrite.resource.ProjectResourceSetSerializer;
-import org.springframework.rewrite.recipes.RewriteRecipeDiscovery;
+import org.springframework.rewrite.RewriteRecipeDiscovery;
 
 import java.nio.file.Path;
 import java.util.List;
