+++ spring-rewrite-commons-launcher/src/main/java/org/springframework/rewrite/RewriteRecipeLauncher.java
@@ -23,7 +23,6 @@ import org.springframework.rewrite.parser.RewriteProjectParsingResult;
 import org.springframework.rewrite.resource.ProjectResourceSet;
 import org.springframework.rewrite.resource.ProjectResourceSetFactory;
 import org.springframework.rewrite.resource.ProjectResourceSetSerializer;
-import org.springframework.rewrite.recipes.RewriteRecipeDiscovery;
 import org.springframework.util.StopWatch;
 
 import java.nio.file.Path;
